Treesame-Commit-Id: b3ade0a63b4180c74766298ddff92c95613c4d0b

+++ app/models/enrollment.rb
@@ -137,11 +137,21 @@ class Enrollment < ActiveRecord::Base
 
   named_scope :active,
               :conditions => ['enrollments.workflow_state != ?', 'deleted']
+
   named_scope :admin,
               :select => 'course_id',
               :joins => :course,
               :conditions => "enrollments.type IN ('TeacherEnrollment','TaEnrollment', 'DesignerEnrollment')
                               AND (courses.workflow_state = 'claimed' OR (enrollments.workflow_state = 'active' and  courses.workflow_state = 'available'))"
+
+  named_scope :of_admin_type,
+              :conditions => "enrollments.type IN ('TeacherEnrollment','TaEnrollment', 'DesignerEnrollment')"
+
+  named_scope :of_instructor_type,
+              :conditions => "enrollments.type IN ('TeacherEnrollment','TaEnrollment')"
+
+  named_scope :of_content_admins,
+              :conditions => "enrollments.type IN ('TeacherEnrollment', 'DesignerEnrollment')"
 
   named_scope :student,
               :select => 'course_id',
@@ -198,6 +208,29 @@ class Enrollment < ActiveRecord::Base
 
   def self.readable_type(type)
     READABLE_TYPES[type] || READABLE_TYPES['StudentEnrollment']
+  end
+
+  SIS_TYPES = {
+      'TeacherEnrollment' => 'teacher',
+      'TaEnrollment' => 'ta',
+      'DesignerEnrollment' => 'designer',
+      'StudentEnrollment' => 'student',
+      'ObserverEnrollment' => 'observer'
+  }
+  def self.sis_type(type)
+    SIS_TYPES[type] || SIS_TYPES['StudentEnrollment']
+  end
+
+  def sis_role
+    self.role_name || Enrollment.sis_type(self.type)
+  end
+
+  def self.valid_types
+    SIS_TYPES.keys
+  end
+
+  def self.valid_type?(type)
+    SIS_TYPES.has_key?(type)
   end
 
   TYPES_WITH_INDEFINITE_ARTICLE = {
@@ -467,13 +500,13 @@ class Enrollment < ActiveRecord::Base
     res
   end
 
-  def accept
-    return false unless invited?
+  def accept(force = false)
+    return false unless force || invited?
     ids = nil
     ids = self.user.dashboard_messages.find_all_by_context_id_and_context_type(self.id, 'Enrollment', :select => "id").map(&:id) if self.user
     Message.delete_all({:id => ids}) if ids && !ids.empty?
     update_attribute(:workflow_state, 'active')
-    user.touch
+    touch_user
   end
 
   workflow do
@@ -757,7 +790,7 @@ class Enrollment < ActiveRecord::Base
   def student?
     false
   end
-  
+
   def fake_student?
     false
   end
